applications/requirements.txt
PyMySQL==1.0.2
Flask==1.1.2
flask_sqlalchemy==2.5.1
sqlalchemy < 1.4.0
sqlalchemy_utils==0.36.8
flask_migrate==2.7.0
flask_script==2.0.6
flask_jwt_extended==4.1.0
cryptography
python-dateutil~=2.8.1
alembic~=1.6.5
redis~=3.5.3